+++ tests/test_services/test_trainer/test_trainer.py
@@ -2,6 +2,7 @@ import pathlib
 import random
 import shutil
 from collections.abc import Callable
+from unittest import mock
 
 import numpy as np
 import pandas as pd
@@ -11,9 +12,14 @@ import yaml
 from phlower.io import PhlowerDirectory, select_snapshot_file
 from phlower.services.trainer import PhlowerTrainer
 from phlower.settings import PhlowerSetting
-from phlower.utils.exceptions import PhlowerRestartTrainingCompletedError
+from phlower.utils.exceptions import (
+    PhlowerNaNDetectedError,
+    PhlowerRestartTrainingCompletedError,
+)
+from phlower.utils.typing import PhlowerHandlerType
 
 _OUTPUT_DIR = pathlib.Path(__file__).parent / "_out"
+_OUTPUT_NAN_DIR = pathlib.Path(__file__).parent / "_out_nan"
 _SETTINGS_DIR = pathlib.Path(__file__).parent / "data"
 
 
@@ -24,6 +30,8 @@ def prepare_sample_preprocessed_files():
     if _OUTPUT_DIR.exists():
         shutil.rmtree(_OUTPUT_DIR)
     _OUTPUT_DIR.mkdir()
+    if _OUTPUT_NAN_DIR.exists():
+        shutil.rmtree(_OUTPUT_NAN_DIR)
 
     base_preprocessed_dir = _OUTPUT_DIR / "preprocessed"
     base_preprocessed_dir.mkdir()
@@ -41,9 +49,19 @@ def prepare_sample_preprocessed_files():
             nodal_initial_u.astype(dtype),
         )
 
-        # nodal_last_u = np.random.rand(n_nodes, 3, 1)
+        nodal_last_u = np.random.rand(n_nodes, 3, 1)
         np.save(
-            preprocessed_dir / "nodal_last_u.npy", nodal_initial_u.astype(dtype)
+            preprocessed_dir / "nodal_last_u.npy", nodal_last_u.astype(dtype)
+        )
+
+        nodal_initial_p = np.random.rand(n_nodes, 1)
+        np.save(
+            preprocessed_dir / "nodal_initial_p.npy",
+            nodal_initial_p.astype(dtype),
+        )
+        nodal_last_p = np.random.rand(n_nodes, 1)
+        np.save(
+            preprocessed_dir / "nodal_last_p.npy", nodal_last_p.astype(dtype)
         )
 
         rng = np.random.default_rng()
@@ -53,6 +71,13 @@ def prepare_sample_preprocessed_files():
         )
 
         (preprocessed_dir / "preprocessed").touch()
+
+    # Prepare data with NaN values
+    shutil.copytree(_OUTPUT_DIR, _OUTPUT_NAN_DIR)
+    u_file = _OUTPUT_NAN_DIR / "preprocessed/case_0/nodal_initial_u.npy"
+    u = np.load(u_file)
+    u[0] = np.nan
+    np.save(u_file, u)
 
 
 @pytest.fixture(scope="module")
@@ -259,3 +284,211 @@ def test__starting_pretrained_state(clean_directories: None):
 
     with pytest.raises(AssertionError):
         np.testing.assert_array_almost_equal(pretrained_losses, org_losses)
+
+
+def test__faster_training_with_on_memory_dataset(
+    simple_training: float,
+    prepare_sample_preprocessed_files: None,
+) -> float:
+    phlower_path = PhlowerDirectory(_OUTPUT_DIR)
+
+    preprocessed_directories = list(
+        phlower_path.find_directory(
+            required_filename="preprocessed", recursive=True
+        )
+    )
+
+    setting = PhlowerSetting.read_yaml(_SETTINGS_DIR / "train_on_memory.yml")
+
+    trainer = PhlowerTrainer.from_setting(setting)
+    output_directory = _OUTPUT_DIR / "model_on_memory"
+    if output_directory.exists():
+        shutil.rmtree(output_directory)
+
+    _ = trainer.train(
+        train_directories=preprocessed_directories,
+        validation_directories=preprocessed_directories,
+        output_directory=output_directory,
+    )
+
+    # load lazy_load csv
+    df_lazy = pd.read_csv(
+        _OUTPUT_DIR / "model/log.csv",
+        header=0,
+        index_col=None,
+        skipinitialspace=True,
+    )
+    lazy_time = df_lazy.loc[:, "elapsed_time"].to_numpy().max()
+
+    df = pd.read_csv(
+        output_directory / "log.csv",
+        header=0,
+        index_col=None,
+        skipinitialspace=True,
+    )
+    on_memory_time = df.loc[:, "elapsed_time"].to_numpy().max()
+
+    # NOTE: the number 1.5 is arbitrary, but it should be larger than 1.0
+    assert lazy_time > on_memory_time * 1.5
+
+
+@pytest.mark.parametrize(
+    "name, allow_overwrite, success",
+    [
+        ("EarlyStopping", True, True),
+        ("EarlyStopping", False, False),
+        ("DummyHandler", True, True),
+        ("DummyHandler", False, True),
+    ],
+)
+def test__attach_hanlder(
+    name: str,
+    allow_overwrite: bool,
+    success: bool,
+):
+    setting = PhlowerSetting.read_yaml(_SETTINGS_DIR / "train.yml")
+    trainer = PhlowerTrainer.from_setting(setting)
+
+    mocked = mock.Mock(spec=PhlowerHandlerType)
+
+    if success:
+        trainer.attach_handler(name, mocked, allow_overwrite=allow_overwrite)
+        assert name in trainer._handlers._handlers
+    else:
+        with pytest.raises(ValueError) as ex:
+            trainer.attach_handler(
+                name, mocked, allow_overwrite=allow_overwrite
+            )
+        assert str(ex.value) == f"Handler named {name} is already attached."
+
+
+# region Test for dumped details in training log
+
+
+def simple_training_with_yaml(
+    setting: PhlowerSetting, output_directory: pathlib.Path
+):
+    phlower_path = PhlowerDirectory(_OUTPUT_DIR)
+
+    preprocessed_directories = list(
+        phlower_path.find_directory(
+            required_filename="preprocessed", recursive=True
+        )
+    )
+
+    trainer = PhlowerTrainer.from_setting(setting)
+    if output_directory.exists():
+        shutil.rmtree(output_directory)
+
+    _ = trainer.train(
+        train_directories=preprocessed_directories,
+        validation_directories=preprocessed_directories,
+        output_directory=output_directory,
+    )
+
+
+@pytest.mark.parametrize(
+    "yaml_file, output_directory",
+    [
+        (
+            _SETTINGS_DIR / "train_for_multiple_targets.yml",
+            _OUTPUT_DIR / "model_for_multiple_targets",
+        ),
+        (
+            _SETTINGS_DIR / "train_for_multiple_targets_with_weights.yml",
+            _OUTPUT_DIR / "model_for_multiple_targets_with_weights",
+        ),
+        (
+            _SETTINGS_DIR / "train_for_multiple_targets_not_evaluating.yml",
+            _OUTPUT_DIR / "model_for_multiple_targets_not_evaluating",
+        ),
+    ],
+)
+def test__dumped_details_training_log(
+    yaml_file: pathlib.Path,
+    output_directory: pathlib.Path,
+    prepare_sample_preprocessed_files: None,
+):
+    setting = PhlowerSetting.read_yaml(yaml_file)
+
+    simple_training_with_yaml(setting, output_directory=output_directory)
+
+    log_csv = output_directory / "log.csv"
+    assert log_csv.exists()
+    df = pd.read_csv(
+        log_csv,
+        header=0,
+        index_col=None,
+        skipinitialspace=True,
+    )
+
+    train_loss = df.loc[:, "train_loss"].to_numpy()
+    train_loss_details = df.filter(regex=r"^details_tr")
+
+    # check train loss details
+    loss_weights = setting.training.loss_setting.name2weight
+    if loss_weights is None:
+        weights = dict.fromkeys(train_loss_details.columns, 1.0)
+    else:
+        weights = {
+            name: loss_weights[name.removeprefix("details_tr/")]
+            for name in train_loss_details.columns
+        }
+
+    _details = np.stack(
+        [train_loss_details.loc[:, k].to_numpy() * weights[k] for k in weights],
+        axis=1,
+    )
+    np.testing.assert_array_almost_equal(
+        train_loss, np.sum(_details, axis=1), decimal=5
+    )
+
+    # check validation loss
+    validation_loss = df.loc[:, "validation_loss"].to_numpy()
+    validation_loss_details = df.filter(regex=r"^details_val")
+    if loss_weights is None:
+        weights = dict.fromkeys(validation_loss_details.columns, 1.0)
+    else:
+        weights = {
+            name: loss_weights[name.removeprefix("details_val/")]
+            for name in validation_loss_details.columns
+        }
+
+    _vl_details = np.stack(
+        [
+            validation_loss_details.loc[:, k].to_numpy() * weights[k]
+            for k in weights
+        ],
+        axis=1,
+    )
+    np.testing.assert_array_almost_equal(
+        validation_loss, np.sum(_vl_details, axis=1), decimal=5
+    )
+
+
+# endregion
+
+
+def test__train_with_raise_nan_detected():
+    phlower_path = PhlowerDirectory(_OUTPUT_NAN_DIR)
+    preprocessed_directories = list(
+        phlower_path.find_directory(
+            required_filename="preprocessed", recursive=True
+        )
+    )
+
+    setting = PhlowerSetting.read_yaml(
+        _SETTINGS_DIR / "handlers/raise_nan_detected_error.yml"
+    )
+
+    trainer = PhlowerTrainer.from_setting(setting)
+    output_directory = _OUTPUT_DIR / "model_w_nan"
+    if output_directory.exists():
+        shutil.rmtree(output_directory)
+
+    with pytest.raises(PhlowerNaNDetectedError):
+        trainer.train(
+            train_directories=preprocessed_directories,
+            validation_directories=preprocessed_directories,
+            output_directory=output_directory,
+        )