+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## UnReleased
+### Added
+* Add `attach_handler` method to `PhlowerTrainer` to add an extra handler at training process.
+* Add `lazy_load` parameter to `TrainingSetting` to load data lazily.
+* When `time_series_length` is -1, `PhlowerGroupModule` determines the time series length automatically from the input data.
+
+### Changed
+* Time series tensor is splitted into each time step when forwarding with `time_series_length` in `PhlowerGroupModule`.
+* Display details of losses at training process.
+
 
 ## [0.2.2] - 2025-06-12